Deprecate OneclickMall and create Oneclick class
Also, deprecate Inscription and Start on OneclickMall
